Encode/decode context stack into an int
When finishing highlighting of a text block, an arbitrary sized
context stack can be encoded into a single integer (as long as
arithmetic overflow doesn't occur).

I think the algorithm is close to optimal, when it comes to how
much info can fit into a single integer.

The encoded value is decoded by dividing it by the number of
possible nested contexts in the surrounding context. The reminder
of the division is the index of the next context on the stack,
and the quotient contains the rest of the stack, recursively.

Indices are encoded as 1-N inclusive, because otherwise the
algorithm would not work with 0 indices.